nested-tabs.js moved to nested-tabs folder
