+++ assets/dev/js/frontend/modules.js
@@ -3,7 +3,7 @@ import Document from './document';
 import StretchElement from './tools/stretch-element';
 import BaseHandler from './handlers/base';
 import SwiperBase from './handlers/base-swiper';
-import NestedTabs from 'elementor-frontend/handlers/nested-tabs';
+import NestedTabs from '../../../../modules/nested-tabs/assets/js/frontend/handlers/nested-tabs';
 
 elementorModules.frontend = {
 	Document,
